+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.48.5] - 2022-03-03
+
+- Added: symfony/polyfill-80 dep
+- Fixed: recipientEmail in submission config type
+- Added: check to submit the right submission
+
 ## [1.48.4] - 2022-02-14
 
 - Fixed: array index issues in php 8+